+++ src/main/kotlin/com/arnyminerz/database/dsl/Transactions.kt
@@ -9,7 +9,7 @@ private const val TRANSACTION_DESCRIPTION_LENGTH = 256
 object Transactions : IntIdTable() {
     val timestamp: Column<String> = varchar("timestamp", DATE_LENGTH)
     val date: Column<String> = varchar("date", DATE_LENGTH)
-    val pricePerUnit: Column<Float> = float("unit_price")
+    val pricePerUnit: Column<Double> = double("unit_price")
     val amount: Column<Int> = integer("amount")
     val description: Column<String> = varchar("description", TRANSACTION_DESCRIPTION_LENGTH)
